Fix the referrer URL if elements are moved inside a nested element (see #6803)
Description
-----------

Fixes #6776

This fixes any issues at the first nesting level, e.g. if you move a content element into an accordion or slider element.

However, it will not fix elements nested beyond the first level, e.g. a content group in a slider element or an accordion in an accordion, as our current referrer management is not designed for such cases and I have not succeeded in adapting the logic.

I would rather revise the referrer management for Contao 5.4 and make it completely client-side.

Commits
-------

3da62fa5 Fix the referrer URL if elements are pasted inside a nested element