Improve the `ConfigureFilesystemPass` performance (see #6848)
Description
-----------

I noticed on my system that calling `vendor/bin/contao-setup` is very slow. I found the culprit to be `ConfigureFilesystemPass::mountAdaptersForSymlinks` - and in particular it happens because of the regex path filtering we use here: 

https://github.com/contao/contao/blob/477ca9e95767212c6c4d507e700654f483778af1/core-bundle/src/DependencyInjection/Compiler/ConfigureFilesystemPass.php#L58

On my system this method can take up to 60 seconds (!) in total. And since this is a compiler pass that is executed _every time_ the kernel initializes the container - and since during `contao-setup` the kernel is initialized multiple times, it consumes _a lot_ of time.

The regex isn't even necessary, since we can just tell the Finder to look in `files/` in the first place. This PR does that.

This improves the execution time of this method from 60000ms to 40ms 🚀

Commits
-------

98bfa5df improve performance of ConfigureFilesystemPass
f11524e5 fix the mount path
2d41f889 catch DirectoryNotFoundException