Do not load the sitemap from cache for authenticated users (see #6832)
Description
-----------

In the Contao back end we have the ability to let the crawler crawl your Contao website(s) with an authorized session of a front end user.

However, under certain circumstances it can happen that the crawler does not retrieve the protected pages from the `/sitemap.xml` route - if it happened to be cached before. This is because the response of the `/sitemap.xml` currently does not set a `Vary` on `Cookie` meaning that despite a session cookie being present the reverse proxy will serve the `/sitemap.xml` always from the cache.

## Reproduction

* In your Contao website create a page and protect it for a member group.
* Also hide the protected page from the menu (and make sure it's not linked to from anywhere else).
* Access the sitemap (`https://example.com/sitemap.xml`) via an unauthorized request (e.g. private browser window) so it is stored in the HTTP cache.
* In the Contao back end go to _Maintenance_ » **Crawler**, enable _Update the search index_, select a member from the select that has access to the previously generated protected page and _Start crawling_.
* Download the debug CSV after the Crawler has finished.

The protected page will not have been indexed, as the Crawler received a cached response from the `SitemapController`.

Commits
-------

1332dab1 add Cookie to Vary header in SitemapController
45251a38 update comment