Do not add duplicate paths in our own Twig filesystem loader (see #6868)
Description
-----------

Fixes #6802, alternative to #6847.

The `ContaoFilesystemLoader` works on internal lists of paths/namespace configurations and template filenames  (that are also cached to reduce the amount of filesystem calls). Currently, the `ContaoFilesystemLoaderWarmer` - a mandatory cache warmer - propagates these lists. 

Calling a cache warmer multiple times must not change the outcome, so there are basically two ways this can be achieved: Either the state must be reset beforehand (= clear the loader and cache) or the calls must effectively be no-ops if the cache was already warmed (= skip). I opted for the latter, because it is really only the `addPath()` calls that change the outcome. With this PR, we are now ignoring duplicate path/namespace directly in the loader. 

*Side note: Our loader extends from the original Twig filesystem loader but it really should not. In the beginning I thought this was a good idea but with the added features, the two are very different now. So, in a future PR I would like to change this and subsequently remove some methods like `prependPath()` that do not make sense in our context and that I also did not touch in this PR. We probably need to discuss if we can/should do this break ("@experimental") or not.*

Commits
-------

b2c3d1c5 skip execution when called multiple times with the same arguments