Correctly output the legend for the security question (see #6947)
Description
-----------

Currently in Contao 4.13 the `<fieldset>` `<legend>` for the security question is missing when using the `member_grouped` template in the registration module. This is because the legend's translation is not pulled from the correct translation key currently. This PR fixes that.

**Note:** this will introduce https://github.com/contao/contao/issues/6937 in Contao 4.13 without the JavaScript fix.

**Note:** this also needs to be merged upstream to Contao 5.3, as there the fieldset does have a legend, but without a translation.

Commits
-------

d6c9794f Fix legend for security question not being output
1c6e5dc5 Fix captcha fieldset check
83dcd1b1 add translation instead
b76a86fd change translation to 'Verification'

Co-authored-by: ausi <martin@auswoeger.com>