Correctly generate record labels with foreign keys (see #6866)
Description
-----------

This fixes a long standing issue that record labels are not correctly generated from foreignKey. I noticed this because of a strange behaviour: I have a custom `label_callback`, but the column view did not show the value my callback returned. This was because the value I returned was used to look up the foreignKey! And for one of 10 records, it actually found something in the foreign tabel (for no good reason) and showed that instead.

Commits
-------

124a9d3c Correctly generate record label with foreignKey
5265c250 Revert the dash placeholder
1dec48b4 Revert the dash placeholder