Revert the changes to the "file uploaded" check (see #7039)
Description
-----------

With file uploads you can (could) control via `'uploaded' => true/false` whether the file will be attached in the email sent by `Contao\Form` or whether it will include a link instead.

However the check for this was changed in https://github.com/contao/contao/pull/4364 from

```php
if ($file['uploaded'] ?? null)
```

to

```php
if (isset($file['uploaded']))
```

meaning that you cannot control this behavior via `true/false` anymore. Now the file will always be inserted as a link only as soon as the `'uploaded'` key is present in the files array.

This PR reverts the check to the previous behavior, so that this can be controlled more easily again.

Commits
-------

420bd8e8 revert the file uploaded check
51b8b2a5 Merge remote-tracking branch 'origin/5.3' into revert-form-upload-change