Also unset the `disable`, `start` and `stop` fields when an admin edits themselves (see #7064)
Description
-----------

Fixes https://github.com/contao/contao/issues/7059

This was actually an inconsistency in all previous Contao versions. The admin user is not allowed to disable itself, because you could disable the only admin account etc. The new user voter prevents that, which correctly disables the toggle operation automatically. But since the fields were still present in the UI, the data could not be stored.

Commits
-------

39b9f2a9 Disallow disable,start,stop fields for admin on itself