Initialize Handorgel on the element (see #7074)
Description
-----------

### Description

- Fixes #7058

If people want to use the new js_accordion.html5 to use handorgel.js for their legacy accordion elements,
it will override a possible body id.

This change addresses a correct initialization on every accordion element and should be BC with current implementations

### Bugfix description

* The initialization has been changed from `document.body` to the parent-element of the toggler
* Also works with other legacy elements and closing them

### Limitations

It didn't work before (won't work with this change either) when using legacy accordions and the new nested accordion elements together. This could be addressed as a technical limitation when using both together.

Commits
-------

426cbe38 Initialize handorgel on the element