Fix the elements check in the `modulewizard.js` script (see #7017)
Description
-----------

It would be nice if can remove this check. It disallows to easily extend the module wizard, as it compares the number of some elements that don't really have an impact on how this widget works. For example, whether you have 2 or 3 select menus in one row does not matter at all, but breaks the widget.

Here is my example of enriched widget with an extra "Language" column:

![CleanShot 2024-03-13 at 12 34 01](https://github.com/contao/contao/assets/193483/3e66f027-d092-46fc-b651-853f18712c43)

Commits
-------

0ff654fb Remove the check elements count check in modulewizard.js
db71a76e Improve the check if row contains all necessary elements
9f172a25 Improve the check if widget is initialized
519e1292 Update core-bundle/assets/scripts/modulewizard.js
ee620802 Check the last element in each row
86ad71b9 Rebuild the assets

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>