Make the member group voter cacheable (see #7089)
Description
-----------

I noticed in the profiler that our `MemberGroupVoter` was asked for a decision for totally unrelated attributes. That's because it did not correctly implement the `CacheableVoterInterface`, it did not limit the attributes it wanted to vote on. This PR fixes that by not using the unnecessary abstract and correctly only `supportsAttribute` on what we want.

Commits
-------

7363ac88 Make the MemberGroupVoter cacheable