Store enum fields in the DCA extractor cache (see #7154)
Description
-----------

**Problem:**
The `DcaExtractor` is parsing DCA data and provides information about any enumeration fields within a DCA (see #6584 for the original feature). However, if the extracted data is cached by the `ContaoCacheWarmer`, the enum configuration is missing.

This results in an incorrect/incomplete configuration and it is not possible to resolve enumerations within a model if the extract is read from the cache.

**Solution:**
This PR adds the enumeration configuration to the cached data in `ContaoCacheWarmer`, making the information available in cached extracts.

**Additional optimisation & Tests**
Manually assembling the extract for the cache does not appear to be the most sustainable solution. However, this is outside the scope of this PR.

I have not made any adjustments to the tests in `ContaoCacheWarmerTest`, as so far only the output of `$this->arrFields` is tested in the extract and any other information (e.g. `arrRelations`, `arrKeys`, ...) is not checked explicitly.

If an addition to the test is desired, I can extend the test case and test data to include a field with an enum configuration.

Commits
-------

407b777b Fix DCA extractor cache for enum fields