Allow hyphens in custom legacy template names (see #7145)
Description
-----------

In https://github.com/contao/contao/pull/731 a deprecation for hyphens in template names was added which turned into an exception in Contao 5.

However it is unclear why this deprecation was added in the first place. The original issue https://github.com/contao/contao/issues/725 was fixed and this fix is still present in Contao 5. There are no actual issues with hyphens in template names, neither in Contao 4 nor in Contao 5.

Besides, the error suggests to use snake case instead, which is semantically incorrect. In legacy templates names the underscore semantically creates template name (sub) groups. If you have a template called `mod_newslist_three-columns` it would technically be semantically incorrect to rename it `mod_newslist_three_columns`.

As discussed on Slack it is unclear whether this deprecation was added in the first place. There are no actual issues with hyphens in template names. However, my guess is we added it because we wanted to change the template naming structure in Contao 5 - but this never happened for legacy PHP templates, as we switched to Twig templates. And Twig templates _do_ have a new structuring scheme.

Therefore this restriction can be removed completely.

_Note:_ I will do a separate PR for 4.13 to remove the deprecation.

Commits
-------

01fdd9a5 remove hyphens restriction for templates names
10a8de95 remove accidental change
6f115fa5 remove test
3f313cce update tests