Invalidate the pagemounts cache in the back end access voter when duplicating a page (see #7102)
Description
-----------

If a user who is not an admin copies a page, the articles (and thus all content) are no longer copied, as it was the case in Contao 4. Instead a new empty article is created.
This is because the check whether the user is allowed to create the articles depends on the pagemountCache in the BackendAccessVoter, which is only set once per request and user. 
As the page does not yet exist when the cache is first set and has not been reset after the page has been created, the check as to whether the user is allowed to access the newly created page fails.

This PR adds a check if a new page has been created before accessing the pagemountCache in the BackendAccessVoter and clears the cache if this is the case.


Steps to reproduce:

1. Login as a user without admin privileges
2. Copy and paste a page that has content elements in its article
3. The original article(s) will not be copied but a new one will be created

Commits
-------

45f2cbb5 Change pagemountCache structure in BackendAccessVoter
5257361d Store true/false in the pagemountsCache
c53c4e10 Merge pull request #1 from aschempp/fix/page-access
f30b8f05 Merge pull request #2 from lukasbableck/5.3
ed3f337c run rector
479dad1c run ecs fix
0c3bba24 Revert "Change pagemountCache structure in BackendAccessVoter"
d031b280 check if page is in pagemountsCache and rebuild cache if not before r…

Co-authored-by: aschempp <andreas.schempp@terminal42.ch>