Add basic support for `hotwired/turbo` in the back end (see #7011)
Description
-----------

This is the first step to add Turbo to the backend. This does not yet enable frames or uses streams at any place but enables drive with instant click/prefetching by default (see below). 

### Status code
The `BackendController` will now return a status code of 422, if a widget was rendered with an error, so that forms still work. For this to work, widgets now set a `_contao_widget_error` attribute on the request in that case, which is evaluated in the controller.

### Assets
There is now a new meta tag `<meta data-turbo-track="reload" content="…">` which contains a hash of the generated stylesheet and script output in the head. If this hash changes, Turbo will issue a full page reload instead of merging. Basically this disables the feature for some pages for now (e.g. the file manager with the Dropzone script added). We can progressively enhance things from here on in follow-up PRs.

### Prefetch
Turbo now - by default - prefetches links on hover. This makes a very nice experience, but we need to disable it for certain things. That is what we are currently doing:

Disable prefetch for…
* … all URLs that contain a `rt` query parameter, so that things like GET requests to delete resources won't be triggered.  
* … everything related to the Sf toolbar.
* … certain menu items (e.g. logout).

We also disable Turbo completely for the `be_login` and `be_password` forms, because it's a completely different structure there.

### User control
If you want to disable Turbo completely, you can do so by setting `Turbo.session.drive = false` in your JS. If you just want to disable prefetch completely, you can add the `<meta name="turbo-prefetch" content="false">` meta tag. 

There is also a fine-grained per-element control using data attributes. See https://turbo.hotwired.dev/handbook/drive#disabling-turbo-drive-on-specific-links-or-forms or https://turbo.hotwired.dev/handbook/drive#prefetching-links-on-hover.

Commits
-------

bbf8962f add basic support for turbo drive and enable it by default
bb713b40 Update core-bundle/assets/backend.js
22642fa8 update built resources
83625b27 update built resources
36837677 do not prefetch anything related to the Sf toolbar
8ba3dcd7 do not prefetch logout link
77f24d7f do not prefetch alerts link
9234d984 merge current 5.x
88a8ca0a automatically reload if any of the combiner scripts/styles changes
4bf0f852 fix tests
116df4d1 update assets
dee9c3f9 do not prefetch debug mode link
a924ba33 gracefully handle missing elements