Render partial tree instead of disabling root trails with breadcrumb (see #7380)
Description
-----------

Fixes https://github.com/contao/contao/issues/7366

As I tried to describe in the call, everything works fine if the tree is rendered from the root. You can see this in the current system by not filtering the tree for `Page_1`. The problem happens when the breadcrumb helper disables `showRootTrails`.

/cc @zoglo

Commits
-------

022fbec6 Render partial tree instead of disabling root trails with breadcrumb