Show a warning if a personal data module allows to change the password (see #7443)
Description
-----------

<img width="830" alt="" src="https://github.com/user-attachments/assets/55defc83-3406-474e-9b7e-baa3af2f8b69">

In contrast to the personal data module, the change password module requires to enter the old password when setting a new password.

Of course, we could adjust the personal data module accordingly (see #7442 for a proof of concept), but changing your password and editing your personal information in the same form seems like an anti-pattern and is not possible on any of the major websites I use. I am therefore against merging #7442. Changing your password should always be done separately from changing your personal data.

Commits
-------

b3d04467 Show a warning if a personal data module allows to change the password
9b3a96e1 Show the warning in the system messages
7c0bbbd7 Only show the warning if the user has access to the module