Add the missing `widget-group` classes (see #7546)
Description
-----------

Follow-up on #7087 and #7229

Currently, if you enable a `submitOnChange` checkbox that enables a subpalette, the CSS will be wrong:

<img width="240" alt="Screenshot 2024-09-22 134228" src="https://github.com/user-attachments/assets/585e9734-5439-48d9-8209-bdc0f22f1709">

\
There will be an additional margin on the `.widget` that was loaded in, due to the no-grid fallback CSS. This margin will go away on a reload:

<img width="239" alt="Screenshot 2024-09-22 134254" src="https://github.com/user-attachments/assets/24e8e182-a7fa-4cba-947a-632648916a52">

\
The reason is that `Backend.toggleSubpalette()` still uses the CSS classes `subpal cf` instead of the new `subpal widget-group`. This PR fixes that (and also changes one other ocurrence where `subpal cf` is still in use).

Commits
-------

6321c3af fix missin widget-group classes