Add a user templates VFS storage (see #7573)
Description
-----------

This PR introduces a new user templates storage: A VFS configuration with a local adapter pointing to `/templates`, that is mounted at `user_templates` and can be accessed by injecting a `VirtualFilesystemInterface $userTemplatesStorage`. The new template studio will use this to create/move/delete user templates.

We can use this in the `ContaoFilesystemLoader` in the future and then make use of the fact, that you can move the target directory around (by mounting). This is not as easy to do, though, because currently Twig (as well as our loader) are built around filesystem paths (see `Twig\Source` or Contao's template hierarchy). Constraining things to a local adapter only could be an option.

Commits
-------

d6efb73a add a user templates VFS storage