Add the data container operations builder (see #7593)
Description
-----------

This adds an `@internal` class `DataContainerOperationsBuilder` which extracts building operations from `DC_Table`/`DataContainer`. The problem this is approaching to solve is using a Twig template for the builder, and allowing the DC to dynamically append or prepend operations to the list defined in the DCA.

Not for global operations (yet)!

Commits
-------

b016a196 Added DataContainerOperationsBuilder
3d8b1adf Fix the callback variable name
5d7e04ec Update core-bundle/src/DataContainer/DataContainerOperationsBuilder.php
383a3437 Reuse the builder class instead of dynamic class
fbde0a7a Rename `prepareButtons()` to `initializeButtons()`
c1e07b8f Fix typo

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>