Hide the picker in "edit all" or clipboard mode (see #7608)
Description
-----------

I noticed we sometimes hide the picker radio/checkbox on select-all, but sometimes don't. For tree mode, the layout is broken when both picker and select-all are active.

I think it would also be wise to hide the picker field if something is in the clipboard, because you don't want multiple operations available at the same time. This would also become confusing/problematic with the upcoming operations submenu.

Related to https://github.com/contao/contao/pull/7593

Commits
-------

3ff4551b Hide picker in edit all or clipboard mode
8ec639c3 Keep the picker value in select or clipboard mode