Fix all deprecations (see #7657)
Description
-----------

This PR fixes all current deprecations in the 5.3 branch. :tada:  For the Twig-related ones, I had to add some checks to distinguish between versions. As soon as we raise our requirements to Twig 3.12, we can remove these.

In the `DynamicExtendsTokenParser`, I completely removed an existing check (debca7ba52af15181c928cebfe8c01e966c71a50). This is now covered by the parser in Twig 3.12, but it is so niche, that I do not think it is worth testing for the version here.

Commits
-------

52ca93db remove the unnecessary "spaceless" filter
e454e066 use FirstClassTwigCallableReady syntax if available
debca7ba remove a check that is now covered by the parser
379387ff simplify test and don't use the deprecated RequestMatcher anymore
a40653d6 remove test case