Check permissions on all operations in the `PermissionCheckingVirtualFilesystem` decorator (see #7637)
Description
-----------

Fixes #7633.

I opted against using a service for now and made the class `@internal`. When we gathered enough experience with the concept, we can still add a factory and make it usable for anyone.

The permission checks only make sense for core virtual filesystems, so the constructor argument was narrowed to `VirtualFilesystem`.

Commits
-------

80285943 check permissions on all VFS operations
0c7b2387 add another invalid path to the tests