Enable `pauseOnMouseEnter` by default (see #7649)
Description
-----------

This enables the [`pauseOnMouseEnter`](https://swiperjs.com/swiper-api#param-pauseOnMouseEnter) Swiper.js setting to improve accessibility by default. As far as I tested this also pauses the slider when a link within a slide is in focus (e.g. when navigating via tab).

Keep in mind though that a dedicated _pause_ button might still be required for accessibility. However, there is no built-in "pause button" functionality in Swiper.js as far as I can see, so this is something that everyone would need to implement themselves (a button that executes `swiper.pause()`).

Commits
-------

5a847a4b Enable pauseOnMouseEnter by default
03d20844 add test