Fix tree rendering with filters and trails (see #7872)
Description
-----------

This is the actually correct fix for https://github.com/contao/contao/pull/7480 (and https://github.com/contao/contao/issues/7366). I have a client which has a `list.sorting.filter` applied by default on the tree, which caused incorrect results, so I had to find a fix.

Trying to explain how it works (copy from whiteboard on my wall 😅):
 - `$dc->root` contains the topmost allowed IDs (`eliminateNestedPages`)
 - `$dc->visibleRootTrails` contains IDs above `$dc->root`
 - `$dc->rootChildren` contains IDs below `$dc->root`
 - when IDs are updated from search results, `$dc->rootChildren` will only contain IDs between `$dc->root` and nested IDs, so we only generate the tree from tomost root until last found node
 - also on search results, `$dc->visibleRootTrails` will contain all IDs that are _between_ nested root nodes, causing their operations not to be rendered (because they are considered root trails).

Initially (`DC_Table::initRoots()`), the arrays will contain the user permissions. `DC_Table::treeView()` will then update `$dc->root` from search results, including the already given permissions. 

Since the original PR was for 4.13, and my client uses 4.13, I had to initially find a fix for their case. This obviously needs to be forward-ported to 5.3 as well, but I guess the same code will run there just fine.

## Some examples from Contao demo

### Admin searching for _slider_, which matches a subpage
Notice how the page `Content elements` does not have operations, because it did not match the search results

<img width="800" alt="Bildschirmfoto 2025-01-14 um 15 20 46" src="https://github.com/user-attachments/assets/e0d188d6-f8c3-43d6-a3ba-fb61f3f9b66f" />

### Admin searching for _elements_, which matches a parent and subpages
Notice how the page `Content elements` now does have operations, because it **did** match the search results, but the _slider_ page is hidden because it does not match

<img width="798" alt="Bildschirmfoto 2025-01-14 um 15 21 02" src="https://github.com/user-attachments/assets/347feaf9-236e-415e-b89f-38ee91998dcb" />

### Admin searching for _detail_, which matches two pages in separate parent nodes
Only the acutal matches have operations

<img width="789" alt="Bildschirmfoto 2025-01-14 um 15 21 35" src="https://github.com/user-attachments/assets/6b6a25e1-f7b1-4e55-ae83-0bd6859f17da" />

### Admin searching for _detail_, while having breadcrumb applied
The _detail_ page of news does not show up because of the breadcrumb menu

<img width="790" alt="Bildschirmfoto 2025-01-14 um 15 25 04" src="https://github.com/user-attachments/assets/16d6da20-c2ec-4922-bc5a-bb88073ba19d" />


### Fancy search for _news|detail_ (regex) 
Operations are now available on the `News` and `News Detail`, as well as `Event Detail`, but not on `Events`!

<img width="880" alt="Bildschirmfoto 2025-01-14 um 15 27 17" src="https://github.com/user-attachments/assets/3de19260-1842-44ab-a697-46cedfe2b67a" />

Commits
-------

8f36e494 Fix tree rendering with filters and trails