Handle missing operations-menu-controller targets within the parent-view header (see #8044)
Description
-----------

### Description

When opening news, you may not always have multiple operations within the header (e. g. the faq bundle does have pasteInto so this issue doesn't happen there).

The operations-menu-controller will try to find the controllerTarget (the button for the menu) which does not exist. This PR checks for that.

![image](https://github.com/user-attachments/assets/232365e5-a2ae-4457-b3ea-021278c3b06e)

Best case would be counting the operations and not registering the controller at all but that is a chicken and egg problem within DC_Table right now as you could modify and add more header operations with e. g. Hooks.

https://github.com/contao/contao/blob/5.5/core-bundle/contao/drivers/DC_Table.php#L4441-#L4480

Commits
-------

89af2494 Handle missing operations-menu-controller targets within parent view …