Use table lock for `contao.cron` service again (see #8053)
Description
-----------

There is an issue with #8001: the file lock would - by default - be a system wide lock, as the lock file would be generated in the system temp directory. So if multiple Contao instances share the same temp directory on the same server, the locking would interfere with each other.

Thus this PR reverts parts of #8001 and implements the proposed table locking with a short lock wait timeout instead. This has the added bonus that there will be more data consistency within `tl_cron_job` itself, in case any other process would want to write something in there.

Commits
-------

3e8d5c23 Use table lock for contao.cron service again
0d80d21b add test