Increase the search delay to 300ms (see #8113)
Description
-----------

With the current delay of 150 ms, Contao sends 8 (!) requests to the server when you enter `demonstration` in the search field at normal typing speed. With a delay of 300 ms, there were only 2 requests and there was no significant decrease in usability.

Commits
-------

e58e7bbc Increase the search delay to 300ms