Restore TinyMCE properly on Safari (see #8127)
Description
-----------

Fixes #8125

This is very annoying. 

When transitioning from a context [A] to [B], Turbo first removes the old DOM (which makes the mutation observer on the controller elements call `disconnect()`), then the new DOM is appended (which makes the mutation observer on the controller elements call `connect()` again). So far everything should be fine. Except it isn't in Safari, where the mutations from removing [A] are not processed until the mutations from [B] are. 

In most cases this isn't an issue, but with tinyMCE binding the editor instance id to the element's id, this leads to a collision in case [A] and [B] both have a tinyMCE controller element with the same id... which is always the case when saving/staying on the same page.

The correct solution would be a fix in Safari (lol), so the best workaround I could come up with is to simply delay the execution of the `connect()` call to when the call stack has emptied and we are basically sure that the disconnect had happened.

I tested it on browserstack but it would be nice if any of you Safari users could test this manually as well.

Commits
-------

fa414a0c fix Safari init issue (fingers crossed)
895b813f use queueMicrotask instead of setTimeout