Add `spellcheck="false"` to the search input field (see #8131)
Description
-----------

Apparently since iOS 11's "Smart Punctuation" feature, the default for quotes are [curly quotes](https://apple.stackexchange.com/questions/299505/ios-11-default-quotation-mark-changed-to-and). So if an iPhone user uses the search and wants to search for a phrase, they would write `“lorem ipsum”` instead of `"lorem ipsum"` - which of course would yield different search results.

There does not seem to be a standardized way to deal with this - however according to [this](https://stackoverflow.com/a/60046254/374996), WebKit will use standard quotes if you add `spellcheck="false"` to an input field.

It may make sense to disable spell checking for a search input field in general - wdyt?

Commits
-------

62ab6ab7 Add spellcheck="false" to search input field