Show errors for failed elements in the back end (see #8181)
Description
-----------

Fixes #7544 for the backend. 

Having a general try/catch for the content element preview makes sense IMO as this would always make it possible to edit the element and fix a potential misconfiguration.

The frontend is not affected for the audio/video element because the error only happens in the `editor_view` block. Image and download elements already failed gracefully in my test. We should not add a general try/catch for the frontend because only the content element itself can know if it should fail gracefully or not.

Commits
-------

9373fd41 Show errors for failed elements in the backend