Also show the edit-all operation if a table is only sortable (see #8172)
Description
-----------

If have a `MODE_PARENT` table that is closed, not editable, not copyable and not deletable. But the records are sortable.
When no operations or global_operations are defined, the sorting operation is correctly added, but the edit-all operation is missing. If I manually add it, sorting works correctly and only the sorting button is correctly shown in select mode.

Commits
-------

086b35a1 Check sorting for edit-all operation