Correctly set the tree root if `showRootTrails` is not enabled (see #8186)
Description
-----------

Yet another fix for https://github.com/contao/contao/pull/8175 / https://github.com/contao/contao/pull/8161

Not urgent for a regular Contao installation. If a tree view does not have `showRootTrails` enabled (which is never the case in the core), `visibleRootTrails` will be empty and `root` might contain IDs that are not root pages. The tree will not be rendered in this case.

This is the case with our [folderpage](https://extensions.contao.org/?p=terminal42/contao-folderpage) extension < v3.3, because it overwrites the breadcrumb but did not implement the new `visibleRoot` filter.

Commits
-------

42366a7c Fix tree root if `showRootTrails` is not enabled
b8526c0f Use the correct table for checking `visibleRoot`