Deprecate the usage of the `typePrefix` property on models (see #8191)
Description
-----------

Currently we set `$model->typePrefix = 'ce_'` / `$model->typePrefix = 'mod_'` in `Controller::getContentElement()`, `getForm()` and `getFrontendModule()`. The purpose of this is so that the `Contao\Hybrid` (`Contao\Form`) element has the correct prefix in the front end - i.e. either `<div class="ce_form">` or `<div class="mod_form">`.

Ideally we should avoid setting dynamic properties on model objects.

This PR changes this so that you can pass the prefix via the constructor instead of the model. If no prefix is passed, it will automatically fall back to `mod_`, if the given model is a `ModuleModel` instance and otherwise it will fall back to `ce_`.

Commits
-------

5e4f2607 Deprecate the typePrefix property on models
41897b53 Also consider Controller::getForm()
d46be36b cs