Fix rendering a wrong template in the back end fallback route (see #8220)
Description
-----------

This is a followup-PR to #8185.

There was still the wrong template rendered in the `BackendController#backendFallback()`. I now added the same fallback as we did in the `PrettyErrorScreenListener` and for the maintenance mode.

With these changes, there were only the old preview toolbar templates left - I therefore also removed them and did the same there, so that the behavior is now consistent for all error templates. As the preview toolbar templates don't have any blocks, extending was not an option, so it is IMHO safe to do so.

Commits
-------

6d530d22 fix backend fallback rendering a wrong template
1ce1357b also drop legacy preview toolbar templates and only render them if th…