Do not use the title attribute in the download content elements (see #8240)
Description
-----------

Fixes #8230

Currently the `MSC.download` translation label is always added to the download's title. However, in the singular `download` content element you can specify a custom title - which would then also be prefixed with `'Download '`. This was not the case in previous Contao versions.

Now, the question is though whether we should apply the `MSC.download` label _only_ when falling back to the file's name for the title attribute - or also if there is a title set in the metadata for the file in the DBAFS? The PR implements the former.

Keep in mind that this change will also affect the download**s** content element (plural), as this is a change in the `_download` component, used by both content elements.

Commits
-------

9c63f8cd Only add MSC.download to fallback title
856a33a1 adjust tests
d33474d4 rename test
dd198e91 revert label change in tests
babcc57c do not use title as text fallback
138d3500 remove title attribute
18dd0556 fix tests
53cae7f9 Merge remote-tracking branch 'origin/5.3' into bugfix/only-apply-down…