Allow absolute URLs in DCA backlinks (see #8322)
Description
-----------

The `config.backlink` DCA key allows to define a custom back link. However, this link is always appended to the current backend route with a query parameter, which means I cannot have a custom route for the link.

This PR fixes that for Contao 5.3. Be aware that it cannot be merged upstream due to changes in regards to the context menu.

The only possible issue of this PR is if someone would have a `/` in their expected query parameters. Not sure that's really an issue though?

Commits
-------

d989c1ea Allow absolute URLs in DCA backlink
22ea0ecc Use the expert's way
f456b2a2 Remove unnecessary parentheses

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>