Deprecate the messenger priority interfaces in favor of the new `#[AsMessage]` attribute (see #8331)
Description
-----------

I think it's time to move on from Symfony 6 and benefit from Symfony 7 features in the new non-LTS versions.

As of Symfony 7.2 you can use the `#[AsMessage()]` attribute on a message to give it a default transport. This has basically the same benefits of our interfaces which we introduced for exactly the purpose of not requiring users to configure the message `routing` in their `config.yaml` whenever they install a new extension that requires background jobs.

However, today, if you want to change the priority of one message only, you would need to completely override the `routing` definition because just adding it would merge the config and it would not have any effect because determining the transport happens from top down.
So say you want to move the priority of `Contao\CoreBundle\Messenger\Message\BackendSearch\ReindexMessage` which is `contao_prio_low` to `normal`. Today you cannot just do this:

```yaml
// config/config.yaml

framework:
        routing:
            'Contao\CoreBundle\Messenger\Message\BackendSearch\ReindexMessage': contao_prio_normal
```

Because this will end up in the final configuration of this:

```yaml
framework:
        routing:
            'Symfony\Component\Mailer\Messenger\SendEmailMessage': contao_prio_high
            'Contao\CoreBundle\Messenger\Message\HighPriorityMessageInterface': contao_prio_high
            'Contao\CoreBundle\Messenger\Message\NormalPriorityMessageInterface': contao_prio_normal
            'Contao\CoreBundle\Messenger\Message\LowPriorityMessageInterface': contao_prio_low
            'Contao\CoreBundle\Messenger\Message\BackendSearch\ReindexMessage': contao_prio_normal
```

So when Symfony Messenger looks at the routing table the `LowPriorityMessageInterface` will still hit first.

With the new attributes, however, we can omit the `routing` configuration in Contao 6. So we'll have just this:

```yaml
framework:
        routing:
            'Symfony\Component\Mailer\Messenger\SendEmailMessage': contao_prio_high
```

Now, you can easily configure it in your `config/config.yaml` because the merged config ends up in

```yaml
framework:
        routing:
            'Symfony\Component\Mailer\Messenger\SendEmailMessage': contao_prio_high
            'Contao\CoreBundle\Messenger\Message\BackendSearch\ReindexMessage': contao_prio_normal
```

And the config wins over the attributes 😊 

Hence, let's deprecate the interfaces now for a more flexible configurability in a future Contao 6. We cannot remove the default routing now as this would be a BC break.

Commits
-------

5538e864 Deprecate the Messenger priority interfaces in favor of the new AsMes…