Always add the operations menu (see #8371)
Description
-----------

Fixes https://github.com/contao/contao/issues/8339

The only reason not to render a submenu is when there are only disabled operations (or operations that are not a link). In all other cases, we do want a submenu to display as context menu. However, if all operations are _primary_ (meaning all menu items are visible), we do not show the controller button. It must exist in the DOM for `accessible-menu` to work though.

Commits
-------

ac2f4f2d Always add the operations-menu