Add a palette helper to support working with the palette manipulator (see #8302)
Description
-----------

I wanted to create this PR for years, now I finally did it. I very often find myself modifying palettes in the `palette_callback` but the `PaletteManipulator` lacks the options to create an instance based on a string and you also cannot ask if there's a legend or a field present. So I always have to resort to `str_contains()` which can of course fail if fields start with the same prefix etc.
So finally, here we go 😅

Commits
-------

718b188e Add more helper methods to PaletteManipulator
c9045850 Add asString() on top
2d8e79f5 Separate PaletteManipulator and Palette
11e1b08e Add Stringable interface
0aaa8e49 Move PaletteManipulator actions to Palette class (#19)
46cc20be Update palette with must-haves
828ac26b Added more tests
18a4b949 Simplify

Co-authored-by: aschempp <andreas.schempp@terminal42.ch>