Prevent deep merging of `contao.messenger.web_worker.transports` (see #8407)
Description
-----------

Fixes #8403

This fix prevents the deep merging, which is not desired, when wanting to override the transports or disable the web worker entirely.

In a default managed edition setup, this value should not be configured, thus getting the default value assigned. Therefore, it should not be a breaking change.

Developers which did rely on the current (and in accordance to the developer documentation, wrong) behaviour, would need to configure the default transports additionally to their currently set ones, like:
`[ 'contao_prio_high', 'contao_prio_normal', 'contao_prio_low', 'another_transport' ]`
(instead of only `[ 'another_transport' ]`).

Commits
-------

5e9a7213 Prevents deep merging of config at contao.messenger.web_worker.transp…

Co-authored-by:  <tom.janke@md-netdesign.de>