Allow search indexers to index protected content (see #8395)
Description
-----------

Currently, if e.g. a content element is protected, it will automatically get wrapped into `` and ``. This is because the core search engine distinguishes only the page access member groups (let's say groups 1, 2 and 3) and if a content element is configured to only one of those groups (2 for example), the contents of this element must not be visible for the other two groups.

However, if you build your own search indexer, things might be different. But currently, there's no way to say you actually want the content of those protected elements to be part of the searchable content.

This PR fixes this by introducing a BC compatible way. We just add another comment `` to identify those elements. When calling `Document::getSearchableContent()` you can now optionally allow those ones to be included when passing `true`. Otherwise, those will get ignored as they always were.

Commits
-------

98a710e9 Allow search indexers to include protected content in searchable content