Set the line height to `normal` for `.tl_listing .tl_left` (see #8353)
Description
-----------

Fixes #8070 by setting the `line-height` back to `normal` as it was also in Contao 5.3.

This however increases the height of a `.tl_listing li` (i.e. the height of one line in the page tree for example) from **35**px to **36**px. The question is do we want to:

1. leave the increased height as it is
2. reduce `--row-padding` from 9px to 8px to compensate - this will result in a smaller height of **34**px
3. reduce `--row-padding` fractionally from 9px to 8.5px to compensate - this will restore the original height of **35**px

For comparison: in Contao 5.3 the `.tl_listing li` height is **30**px.

Commits
-------

0fc61156 Do not set line-height for .tl_listing .tl_left
097c4fc6 add min-height