Remove the `title` attribute from the breadcrumb elements (see #8439)
Description
-----------

Just as in https://github.com/contao/contao/pull/7839, https://github.com/contao/contao/pull/8240 this removes the `title` attribute from the breadcrumb elements. The `title` would be the same as the link name, if no `pageTitle` was set - and even if the `pageTitle` is set, the `title` should not be used. And it likely also does not make sense to use the meta `pageTitle` as `aria-label`.

Commits
-------

0ec82e69 Remove title attribute from breadcrumbs