Fix side-effects in `Document::getContentCrawler()` (see #8455)
Description
-----------

Currently, if multiple listeners (or just consumers) of `Document::getContentCrawler()` work with the `Crawler` instance and e.g. add or remove nodes, they will affect each other which is never the desired outcome.

We have to make sure, those instances always represent the original document body.

The easiest fix would've been to not re-use any object but just simply always return `new Crawler($this->body)` but this would mean that the HTML gets parsed for every single call again. So I tried to be smart about this.

Commits
-------

0cc77fd8 Fixed side-effects in Document::getContentCrawler()