Restructure the `flexible` theme and rewrite it using PostCSS (see #8469)
Description
-----------

### Description

* supersedes #7496 

As discussed in the dev meeting in 2024, we want to go native and have decided to use PostCSS for autoprefixing.

This PR aims to split the flexible theme into more readable and smaller parts. Please note that this is by no means the final order of CSS files, as we need to ensure that the output after building the assets remains exactly the same (1:1).

#### Current Structure

Basically, it follows a 7-1 Sass guideline structure. Note that some files and CSS are NOT currently in their ideal places, to maintain the same CSS output after building the assets.

* :open_file_folder: abstracts
  * helpers suchs as `.invisible`, `.clear` etc.
* :open_file_folder: base
  * base includes that are being used globally
* :open_file_folder: components
  * component styles
* :open_file_folder: layout 
  * backend layout, list views etc.
* :open_file_folder: pages
  * includes the other css files that are specifically used in different views
* :open_file_folder: styles
  * specific styles that enhance
* :open_file_folder: vendors
  * vendor styles such as mootools components, tinymce, ace, dropzone, etc.
* :page_facing_up: `main.pcss` _(main include)_
  
### Review / Testing

As discussed, rewriting this should end up in the same css output that we have right now. Building the assets will create the same output right now.

I did "disable" / rewrite the postcss.config for the sake of testing and matching the same css output, it can be reenabled once reviewed.

![image](https://github.com/user-attachments/assets/7b2e6eba-8f41-4441-82e8-0ed0f99c5317)
![image](https://github.com/user-attachments/assets/5dd6d790-7c9d-4440-9d4c-e4eadcb2933b)

Commits
-------

faad68a9 Only use `postcss-preset-env` for the modern css
b0504a8b Move the import files into their respective locations
b32f73b5 Add the split up css files