Compress the serialized search document (see #8473)
Description
-----------

This would be an improvement for https://github.com/contao/contao/issues/8197.

Currently, if you have a 1 MB HTML document, we would write 1 MB to the database. This simple compression logic would bring that down considerably.
I guess I would be okay doing that but not sure we want it and if so, 5.3 or 5.6 - marking up for discussion.

Commits
-------

59a28795 Compress serialized search document
bb3d9309 Test BC
b46a5f3b Use base64 to encode null bytes