Remove `searchResultLimit` from Choices.js  (see #8502)
Description
-----------

I was getting confused in Contao 5.5 why I couldn't find certain entries in large selects, like the `type` of modules or content elements, unless I typed in more letters. For example, if you search for _reader_, the _Event **reader**_ module will not be in the search results. Or if you search just for _p_, the _Change **p**assword_ module will not be in the search results, etc.

The reason for this is that _Choices.js_ has a default `searchResultLimit` of just _4_. This default setting does not make any sense to me, because the default for `renderChoiceLimit` is already `-1` (i.e. disabled). So when _not_ searching, you see _all_ entries, but when you enter a search - which would already filter it down to less entries than before - it is suddenly limited to only 4.

This PR disables the `searchResultLimit`.

Commits
-------

19393ef9 Remove Choices.js searchResultLimit
b51b2ef6 Use -1