Add a general `From` override for the `ContaoMailer` (see #8495)
Description
-----------

In https://github.com/contao/contao/pull/1830 we added the ability to be able to select different transports for different website roots (or different forms or newsletter channels) - with the additional ability to be able to override the `From` address per transport, as a specific transport might use an SMTP server with authentication and thus only a specific `From` address is allowed by the SMTP server. This allows you to have a different administrator email set in the system config while still sending all emails with the appropriate `From` address dictated by the SMTP server.

However, if you do not need multiple transports, but still need to be able to force the `From` address due to the aforementioned limitation, this is currently cumbersome. 

1. You need to specifically define a `mailer.transport`
2. And then also specifically define that transport's `From` address via `contao.mailer.transports.*.from` 
3. _And then_ also select that transport in the website root of your website. 

And then this will only apply to emails generated by a request to the respective front end. Any other system generated emails (via a Cronjob for example) will not be affected.

This PR adds a simple

```yaml
# config/config.yaml
contao:
    mailer:
        override_from: Override From <override@example.com>
```

config which will be applied to _any_ email sent via the `ContaoMailer` (Symfony Mailer) - unless of course a specific `X-Transport` is chosen for which a `from` address has been defined, then that `From` address will of course still be used.

This allows you to only configure the SMTP server via the normal `MAILER_DSN` environment variable and then override the `From` address via the config, without having to specifically define and select any mailer transports.

_Note:_ this will not affect Symfony's `mailer:test` command, as this command bypasses the `mailer` service.

Commits
-------

dd74cdd3 Add contao.mailer.override_from config
62dd6e8a check for null
b3cfaec7 Adjust test