Simplify the positioning of the operations menu (see #8518)
Description
-----------

### Description

As discussed, when using `position: fixed`, an open operations menu will scroll within the view. Since we do not want some wonky layout shift with the body with `overflow-y: hidden` (due the appearing/disappearing scrollbar), it's the expected behavior of how `fixed` works.

This PR simplifies the positioning of the operations menu in the case where the trigger for opening it was **NOT** the `contextmenu` event, thus always positioning it absolute when opening it through the controllerTarget / the button. We also do not need any offset anymore, the `2px` are the ones from the `border-width` and are always accounted.

This also fixes #8477 in the upcoming version (5.6) due to the nature of positioning.

Commits
-------

ff6712ef Simplify the positioning of the operations-menu
683f1d68 Run linter
08893456 Rewrite the operations menu to be absolute on the `contextmenu` event…
53348c84 Wrap `cte_preview` within parent views
d46b2e4e Typo (:
4df030a1 Also remove the unnecessary `linear-gradient` on `hover` within the l…
b0244f11 Calculate from the `offsetParent`