Show the drag handle in `MODE_PARENT` again (see #8575)
Description
-----------

Fixes #8551

The issue was that the `<button class="drag-handle">` and its `<li>` parent was always set to `display: none` by the `sortable-controller`. It checks whether the total amount of drag handles equals 1, and if not, hides the drag handle in the operations. 

_But_ there will always be at least two drag handles per row: one for the light mode and one for the dark mode. Thus the drag handle is never shown.

This PR fixes that by adjusting the JavaScript accordingly.

\
_Note:_ this is subject to change, once https://github.com/contao/contao/issues/8155 gets implemented.

Commits
-------

971a0095 Show drag handle in MODE_PARENT again
444d5690 remove PHPunit result cache
95cf698b Update core-bundle/assets/controllers/sortable-controller.js
2035e26e rebuild assets
ca792a9c update comment