Fix the title tag in Twig layouts (see #8682)
Description
-----------

Fixes #8679

This adds a new `block` to our `_html_boilerplate` so that we can only adjust the `<title>`. For our Contao pages we want to restore the original _Page name - Root page name_ title style, which this PR adds to the `regular.html.twig`.

The new block then allows you to adjust the title style to your needs, e.g. if you want to have a `|` as the separator:

```twig
{# templates/page/regular.html.twig #}
{% extends "@Contao/page/regular.html.twig" %}

{% block title %}
    {% set title = head.title|default %}
    {% if title and contao.page.rootPageTitle|default %}
        {% set title = title ~ ' | ' ~ contao.page.rootPageTitle %}
    {% endif %}
    {{ parent() }}
{% endblock %}
```

The title is taken from the `HtmlHeadBag` which the `AbstractPageLayoutController` passes to the template in the `head` variable.

Commits
-------

4a98513e Fix title tag in Twig layouts and add block
f34e3737 set title variable instead
b220843c add is not empty check
085d40a2 simplify check