Remove the debug markup from the Combiner (see #8695)
Description
-----------

This is an attempt to fix #8686 by just removing the whole `Combiner::getDebugMarkup()` functionality. Instead we just disable `combineScripts` in debug mode which ends up with the same result as the `getDebugMarkup()` did as it outputs seperate script and style tags.

So for the core itself the behavior is the same, but if an extension is using `Combiner::getCombinedFile()` in debug mode they get a combined file back now. I think this is fine as it does not break anything.

For clarification: `Controller::replaceDynamicScriptTags()` is the only place we are using `Combiner::getCombinedFile()`.

Commits
-------

bf642a92 Remove debug markup from Combiner