Enable the `appendGroupInSearch` option for Choices (see #8696)
Description
-----------

### Description

This enables the previously not working option `appendGroupInSearch` within choices as I did plan on using it when replacing choices with chosen (but it did not work). This will be addressed in the next release of choices (see https://github.com/Choices-js/Choices/issues/1325 and https://github.com/Choices-js/Choices/pull/1324).

#### Context

When using `moo_chosen` in Contao 5.3, the optgroups were shown whilst searching for an entry, thus making clear where e. g. a content-element belonged to. This is now particularly cumbersome when having elements that may belong to another optgroup having the same name (or overriden templates from bundles).

**Contao 5.3**
<img width="435" height="282" alt="image" src="https://github.com/user-attachments/assets/e9e66a28-28ad-46b9-8276-b6011239a8a5" />

**Right now**
<img width="456" height="186" alt="image" src="https://github.com/user-attachments/assets/cf607710-4f57-4970-ad19-4bb6f986cc31" />

**After enabling this and upgrading to choices 11.2**
<img width="407" height="190" alt="image" src="https://github.com/user-attachments/assets/f447245d-ca6b-4ee6-9951-102f62e8faf5" />


#### Leading whitespace

The leading whitespace seems to be an issue on our side 
<img width="358" height="119" alt="image" src="https://github.com/user-attachments/assets/7226e19b-0217-4380-acd7-985b4053bbf1" />

https://github.com/contao/contao/blob/a0110932e9e6aba510887fc5293fa38fe26f921e/core-bundle/contao/widgets/SelectMenu.php#L186

<img width="287" height="183" alt="image" src="https://github.com/user-attachments/assets/58bcf643-a7e1-406c-a4eb-42014ed28d40" />

Commits
-------

3ed7e16a Enable the `appendGroupInSearch` option for `Choices`
b5952874 Remove the `nbsp` from the opt groups
94145a0a Revert optgroup whitespace on image size