Fix the search indexer migration when migrating from Contao 4.13 to 5.6 (see #8707)
Description
-----------

When updating from Contao 4.13 to 5.6 the following errors will occur during the database migration:

```
  [PDOException (22007)]
  SQLSTATE[22007]: Invalid datetime format: 1292 Truncated incorrect DECIMAL value: ''  


Exception trace:
  at vendor\doctrine\dbal\src\Driver\PDO\Connection.php:33
 PDO->exec() at vendor\doctrine\dbal\src\Driver\PDO\Connection.php:33
 Doctrine\DBAL\Driver\PDO\Connection->exec() at vendor\doctrine\dbal\src\Connection.php:1216
 Doctrine\DBAL\Connection->executeStatement() at vendor\contao\core-bundle\src\Migration\Version505\OverwriteMetaMigration.php:72
```
```
 [PDOException (22007)]
  SQLSTATE[22007]: Invalid datetime format: 1292 Truncated incorrect DECIMAL value: ''  


Exception trace:
  at vendor\doctrine\dbal\src\Driver\PDO\Connection.php:33
 PDO->exec() at vendor\doctrine\dbal\src\Driver\PDO\Connection.php:33
 Doctrine\DBAL\Driver\PDO\Connection->exec() at vendor\doctrine\dbal\src\Connection.php:1216
 Doctrine\DBAL\Connection->executeStatement() at vendor\contao\core-bundle\src\Migration\Version506\SearchIndexerSettingsMigration.php:46
```

The first error occurs because the `OverwriteMetaMigration` happens to be executed _before_ the `BooleanFieldsMigration`. Thus the `overwriteMeta = 1` query results in the (cryptic) SQL error, since `overwriteMeta` will still be a `CHAR(1)` at this point.

The second error occurs because we are using `noSearch = 1` - again because `noSearch` will still be a `CHAR(1)` at this point - but it will never be migrated to a `TINYINT` in this case, since it does not exist in the DCA anymore.

This PR fixes the former by making sure the `BooleanFieldsMigration` runs first and the latter by using `noSearch = '1'` - which is also safe for `TINYINT`.

Commits
-------

44dbfad1 Fix 4.13 to 5.6 migration errors
5d406f7c add comments
b3a0fd5e revert changes to the OverwriteMetaMigration (see #8708)