Use a 16:9 aspect ratio in translations (see #8755)
Description
-----------

Super minor issue - but I find the 4:3 aspect ratio in our video player size description somewhat outdated. But feel free to disregard and close if it's not worth the effort of then also changing this in all languages 🙃

Though this will then match the default size of our YouTube and Vimeo players.

Commits
-------

4e360744 Use 16:9 aspect ratio in translations