Make a job optional for the back end search reindex (see #8766)
Description
-----------

Currently, you'll see all backend search re-index jobs in the new jobs framework which will be a lot.
Basically, every edit of any content triggers a re-index which is correct because the contents might have changed. So updating the search index is correct but showing those jobs in the jobs framework is of no real benefit because things will happen mostly instantaneous (indexing/updating the index with e.g. one content element will be super quick, why bother showing any progress in the jobs framework).

Let's limit this to the ones executed in the backend maintenance mode only because those are the ones that take a while.

I've decided to add a new feature on the `ReindexConfig` as otherwise every dev would have to inject both, the `BackendSearch` as well as the `Jobs` service in order to trigger re-indexing **with** a job. Like this, it's not needed and the DX is better as you can just do `withRequireJob(true)` and the `BackendSearch` will then create it for you.

Commits
-------

910bfee9 Make job optional for backend search reindex