Fix the Stimulus color picker preview (see #8768)
Description
-----------

### Description

- fixes #8504

The root cause of the issue is stripping the rhombus to comply to our own input field with eval color and initializing the default with `ffffff`. The used library expects hex colors with the rhombus prepended (see value on [input type color](https://developer.mozilla.org/en-US/docs/Web/HTML/Reference/Elements/input/color)), in this specific case it did work without the rhombus (`#`) but not for `ffffff` so allowing other values is more of a feature than a bug.

#### Explanation for the fix

Since our color picker allows different values than hex (depending on eval), I check whether the value matches the HEX format first, then _temporarily_ prepend a `#` for the `connect` method that builds the picker as seen here:
https://github.com/stimulus-components/stimulus-components/blob/c44e23a93c469ed41b4d1088fa4639b94f639443/components/color-picker/src/index.ts#L23-L30

I simply strip the attached rhombus again afterwards

#### Disclaimer

I know there is the `inputTargetConnected` that I could have used for that but there was no need as we simply have to handle it just right before the init of the `Pickr` library.

Commits
-------

5be0c660 Fix color picker init with `ffffff` as a default
ce3324af Lint
b492c856 Consider short hex values as well / Also rebuild the assets