Re-add `webpack-dev-server` and configure hot module replacement (see #8494)
Description
-----------

### Description

The release of webpack-encore v5 dropped the webpack-dev-server as a dependency and is considered a BC Break, basically why our current config doesn't work:

https://github.com/contao/contao/pull/8371#discussion_r2113350606

> yes it is unrelated, but I noticed it does not work at all 🙃


* https://github.com/symfony/webpack-encore/releases/tag/v5.0.0
* See https://github.com/symfony/webpack-encore/pull/1318
* See https://github.com/symfony/webpack-encore/pull/1336

This PR adds `webpack-dev-server` in v5 as a dependency and adds some configuration to enable hot module reload.

![image](https://github.com/user-attachments/assets/860f5a7d-9dd5-4cbc-bf9e-6abe0834a347)

___

### Running one devServer for both configurations

The current configuration within the `webpack.config.js` checks for the CSS files within the following two paths:

* core-bundle/assets/styles/*
* core-bundle/contao/themes/flexible/styles/*

Changes that are detected in these styles will trigger the reload of the page, thus making it easier for us to work with our assets.

As of right now we have 2 Encore configurations, thus kicking out the devServer from the second configuration (as you would have to tie it to another port). The first configuration actually embeds the output path of the backend theme `flexible` so it works both ways.

___

### Should this be a Bugfix and backported to 5.3?

The dev-server has been broken since changing to webpack-encore ^v5 but this is more for DX when working on contao. Backporting this change could be done in a separate PR but I don't think it's really necessary.

Commits
-------

af290cf7 Re-Add webpack-dev-server and configure HMR
eb151492 Rename the backend.js asset
921ef56f Also consider the `themes/flexible` icons and fonts for HMR
2825107e Trigger pcss compilation on imported files
8add0747 Match the webpack configuration for the new IconMinimizer feature
65d6e704 Watch the `core-bundle/contao` directory for template changes as well
c4933afb Also add the new `contao-backend` asset to all occurrences
e84e01ab Update the package.lock
a7b6b8f6 Update package-lock
5545373b Revert the `contao-backend` entry and update the public path in DevSe…
e6a09647 Update `package-lock.json`
46af7a7a Revert and rebuild the assets
4c850230 Simplify webpack configuration
9818a8d1 Update webpack configuration
d0517850 Remove the trailing slashes and uncomment the live reload for HMR