Do not hardcode the file preview (see #8819)
Description
-----------

The file preview is currently shown if the user does not have access to all of the important part fields, which makes no sense. Additionally, it did render a `placeholder.svg` if the image could not be rendered, instead of just not showing a preview at all.

The additional advantage of the new DCA field is that you can manually choose to display the preview in edit-all mode or only show/edit the file name.

Commits
-------

fa3ce429 Do not hardcode the file preview
05e93508 Show exception message when image cannot be created