Do not check the publishing state for content URL results (see #8936)
Description
-----------

Fixes #8915

When generating content URLs for news, events etc, we absolutely do know the target page. If that target page is unpublished, the routing currently fails. That does not make much sense though, because you do want to get to the correct page, but want to get a 404 if preview is not enabled. Or be able to enable preview if you are on the preview entry point.

Commits
-------

7d98c56d Do not check publishing state for content URL results