Add templates for `DC_Table::edit()` (see #8981)
Description
-----------

Followup to #8975 

This PR…
* moves the HTML out of the `edit()` method.
* adds a simple utility class `ArrayTree` as a data structure (more on this below).

## Review notes
### AJAX mode
There are two control flows in the method: One, that returns the full edit form and one ("AJAX mode"), where we only output a specific subtree of widget groups. For this a variable `$return` was used to step by step build the full form and a separate structure to build the AJAX version. When the condition to output the AJAX version was met (=  `Environment::get('isAjaxRequest')` plus the current widget group's id matching a given `$ajaxId`), the `$return` was simply ignored. 

Reading the old code, you can see, that `$blnAjax` can only be `true` if `Environment::get('isAjaxRequest')` is true and as soon this happens, `$return` will be ignored and the method will early return once `[EOF]` is hit with the current level's id. 

Important to note, is the actual content - the `row()` as well as the structure of sub palettes - must of course be the same. This now nicely translates to the following structure of templates:
* A `backend/data_container/table/edit.html.twig` for the full form,
* a `backend/data_container/table/edit_ajax.html.twig`, which is the part rendered in AJAX mode
* and a `backend/data_container/table/_widget_groups.html.twig` that is used by both and renders a given tree.

### Tree
Generating a tree data structure for rendering is always a bit cumbersome when the procedure to do so is imperative. So for example "add that content", "go one level deeper", add another bit", "go one level up", … I therefore decided to add a simple helper structure `ArrayTree` that easily allows doing this:

```php
$tree = new ArrayTree();

$tree->addContent('foo');
$tree->enterChildNode();
$tree->addContent('bar');
$tree->up();
$tree->addContent('baz');

$tree->toArray(); // ['foo', ['bar'], 'baz']
```

(see the unit tests for more details)

We now use this data structure to store the generated tree of widget groups for each box. As it itself is an `ArrayIterator`, traversing it is very neat in the template. I'm likely going to use the same structure for the Template Studio navigation tree in the future. I marked it `@internal` for now, so that we can change the usage if we need to.

/cc @diekatrin

Commits
-------

ab369c57 add a utility to iteratively create tree an array tree structure
a7c9c0b9 extract HTML from edit into templates
5936af1d make PHPStan happy
1e62ae7f generate form buttons, then messages, then the back button
c7da071d generate the messages last and make the form buttons part of the othe…
917bcd42 remove wrong parentheses
49566d75 always pass buttons object to the templates