Add deferred execution for Twig layouts (see #8774)
Description
-----------

Fixes #8870

See #8705 ('The response context headbag is generated before the modules…'), #8709

## The issue
Due to the nature of a page layout, we are now doing/outputting things that affect the response context in the same template as the response context itself is output. For this to work, the latter must be deferred. 

**Example (shortened):**
```twig
<head>
  {% for element in response_context.end_of_head %}
    {{ element|raw }}
  {% endfor %}
</head>
<body>
  {# Content element that adds a style tag to the head #}
  {{ content_element(5) }}
  
  {% add to head %}
    <meta name="foo" content="bar">
  {% endadd %}
<body>
```

In the above example (= state before this PR), the head would **not** contain the respective style and meta tag.

## The solution: deferred execution
The PR introduces a `defer` tag for Twig and a renderer handling it. The above example adjusted would now produce the expected output:
```twig
<head>
  {% defer %}
    {% for element in response_context.end_of_head %}
      {{ element|raw }}
    {% endfor %}
  {% enddefer %}
</head>
[…]
```

### How it works
The `defer` tag is handled similarly to the `block` tag: When parsed, a new block context is generated (= a new block method in the resulting PHP template file). But instead of the usual block reference node (= the call to the block method), we output a `DeferredBlockReferenceNode` which outputs a `DeferredStringable` - a stringable object that holds a closure to the block method call, that only evaluates in the `__toString()` method.

When now using streaming to get the chunks…
```php
$chunks = $twig->load('template.html.twig')->stream($context); // \Generator<string|\Stringable>
```
… we can concatenate them ourselves and thereby evaluating the `DeferredStringable` objects last. It would be nice if we could do this in the environment's render method instead of having a second render service but Twig does not support decorating the environment. 

### Considerations
* Why not use https://github.com/rybakit/twig-deferred-extension? Unfortunately, the extension's design isn't compatible with Twig's `use_yield` mode which will be mandatory in Twig 4 (and its currently broken anyways). This is also why we need a specific renderer and cannot implement it in the compiler step.
* Why not make `defer` a block attribute (`{% block foo defer %}`)? We could still do this but I decided against it, because then new edge cases will appear - like: What should happen when using the `block()` function or when reusing the block in one of the multiple ways?

Commits
-------

5a97e355 add a new defer tag and renderer to Twig
14fa1e25 suppress virtual defer blocks in the inspector
bde25c67 handle deferred blocks in the abstract page layout
0e54c932 adjust the default layout (defer tag)
7f3888de do not wrap object in closure (by reference anyways)
7e048d83 fix param tag
b64ac43b raise Twig requirement to 3.18
9a92a986 make phpstan happy
f84d2674 also raise the twig-bridge version (see https://github.com/symfony/sy…
2dda574f improve generation of unique block name
bcaeb8e0 use xxh3 hash
88ecc155 reset static properties after testing
949ee5dc move outputting the title into the defer block
65d4ba2c move outputting body tag into a defer block
e6e9cc82 Merge remote-tracking branch 'origin/5.6' into bugfix/twig-layout/defer
dc51e538 fix test
7fa1cee6 add assets context to the service locator
1db375e3 Merge remote-tracking branch 'origin/5.6' into bugfix/twig-layout/defer
faa0f5a5 drop duplicates
a44304fc fix error from merge