Add progress for the back end search jobs (see #9013)
Description
-----------

This introduces progress for the back end search jobs. I noticed that there are of course cases where you cannot determine the progress because there is no total known upfront. So far, we do not have such a case.
Maybe in that case, it would be better to just show a `pending` state without progress?

Anyway, the way it is implemented for now, I think it should accept a `null` for `$total` which does some random fun in order to show something is happening :)

Commits
-------

2e7b2d8a Update job progress for backend search
5d6fc9c5 Use a different logarithmic algorithm to indicate monotonic progress