Move the favorites button to the breadcrumb (see #8834)
Description
-----------

This moves the favorite icon from the orange header bar to the breadcrumb. This both saves room in the header and makes more sense, because the current breadcrumb position (= current URL) is what you are adding to the favorites.

<img width="484" height="175" alt="Bildschirmfoto 2025-09-10 um 13 25 46" src="https://github.com/user-attachments/assets/e0cf370e-00fb-4bd7-9ea1-486b994753ae" />


Currently darft because I'm not yet happy with the controller behavior.

Commits
-------

65954439 Move the favorites button to the breadcrumb
15b333df Improve positioning
41b8a966 Rewrite the controller to work without response exceptions
69f97c34 CI
38e1e241 CI
2ecadd40 Tests
21e486d8 CI
142d9a78 CI
0c35af1b Adjust the SVG colors

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>