Implement `multipleFiles` option in file upload form field (see #8890)
Description
-----------

This PR implements uploading multiple files in a single file upload form field.
If enabled the `multiple` HTML attribute is added to the field and `[]` has to be appended to the name attribute. 
The final field value is an array of the uploaded files.
The value of upload fields without the `multipleFiles` option enabled is the same as before, so no breaking change.
If `storeValues` is enabled in the form, the files will be stored as a serialized array in the corresponding database column.

Commits
-------

79aca271 Implement multipleFiles option in upload form field