Fix `isHidden()` for `ContentProxy` (see #8997)
Description
-----------

Fixes #8013 

This fixes the aforementioned issue as discussed in #8976. Since `Contao\ContentProxy` does not load the model's data onto itself and has it's own `__get()` method which only accesses the `templateProperties` if present, `$this->invisible` (etc.) in the `ContentElement::isHidden()` method simply do not exist and thus fragments were always rendered when retrieved via `Controller::getContentElement()`.

This affects the `content_element()` Twig function, the `{{insert_content::*}}` Insert Tag as well as the `alias` content element.

This PR fixes this by populating `$this->arrData` of the `ContentProxy` instance (just as regular `ContentElement` instances would) and then allowing access to that data via `__get()`. If custom data is set via `templateProperties`, that data still has priority.

Commits
-------

64ea5a74 Fix isHidden for ContentProxy
7e31a4f9 Do not change __set()