Re-add the `flippedState` for the Choices dropdown (see #9043)
Description
-----------

### Description

- fixes #9039 

I've been seeing DomToken errors in the console quite a few times but never got to debugging the reason why.
<img width="786" height="185" alt="image" src="https://github.com/user-attachments/assets/9a7f1f8f-a6fe-4e51-9e09-8543808b9f4c" />

This is because of trying to add an empty "flippedState" class within choices, thus leading to an error (as it can't apply an empty string).

This PR reverts the behavior and also allows to flip the select drop down based on how far down it is:
<img width="630" height="522" alt="image" src="https://github.com/user-attachments/assets/91182c91-cc20-47ab-9ea9-9bcc473667a4" />

/cc @leofeyer 

If we want to introduce the intended behavior, we would also have to change the border-radius for 5.7 based on the `is-flipped` class. If the behavior is not intended (I slighty remember talking about this last year), then we could use any other class.

### Reproducer

You can reproduce this in the current demo and scrolling as high up that the e.g. the layout module selection can be barely seen. 
<img width="1193" height="289" alt="image" src="https://github.com/user-attachments/assets/98b30db3-fb67-4727-8cb5-5794335b3cd8" />

If you scroll further down so the drop down can be fully shown, it will autofocus correctly:
<img width="606" height="459" alt="image" src="https://github.com/user-attachments/assets/ef94d744-fdf3-453a-a4c9-fb1c88a46292" />

Commits
-------

20a70e4b Readd the `flippedState` for the choices dropdown
c6a1db51 Downgrade dependencies