Integrate the message bus in the jobs framework for better DX (see #9083)
Description
-----------

The Jobs framework has been designed to work independently of the Messenger integration. This is good because as you can see in our current backend search implementation, the job integration can be optional. Or you might want to integrate a job with your cronjobs for example.

However, the majority of services interacting with the jobs framework will want to dispatch messages for async processing with the Messenger. Right now, you always have to inject both, the jobs as well as the message bus services if you want to do that. 
This PR improves DX for this use case. You now only have to inject the jobs service 😊 
I'm also planning to use that in https://github.com/contao/contao/pull/8826.

Commits
-------

ad184f4d Integrate message bus in jobs framework for better DX