Improve the performance of the table data container provider (see #9072)
Description
-----------

Currently, we're executing a `SELECT *` on every DCA when indexing for the backend search. This is of course nonsense because it will cause a ton of useless data being retrieved from the database.
This will both, speed up the indexing as well as reduce the data transfer from the db.

Commits
-------

3d79840f Improve performance of TableDataContainerProvider
3ecd8215 Adjust the comment