Add a "contao-main" Turbo frame and apply filters to it (see #8971)
Description
-----------

This introduces a Turbo `contao-main` frame. Used only for the DC filters for now but we can improve on that.
The advantage is that we don't have to generate the header and navigation on those requests. I did a benchmark and in my simple Contao 5 demo case which consists of basically no content whatsoever generating the main navigation and the header takes about 25 ms. Or in other words, this PR saves you 25 ms on every contao-main Turbo frame request 🚀

Commits
-------

f065355e Implement a Turbo frame for main area
6416edf0 Restore previous behavior