Improve how the `ide-twig.json` file is generated (see #8200)
Description
-----------

Implements #8051

This PR improves the way the `ide-twig.json` IDE helper file is generated. Currently we are only creating the file during cache warmup in the dev environment in an *optional* cache warmer. This is unfortunate because you typically do not explicitly warm the dev cache. We also dumped the file to `var/cache` which is semantically wrong.

Now the file gets dumped to `var/contao-twig` in a *mandatory* cache warmer (still only in dev). In order to use it you can un-exclude the above directory in your IDE.

I also added a new command `contao:dump-twig-ide-file` that dumps the file to a given location in your project dir. This way it will be included by default and can be checked in to version control. In this case, you own the file, so in order to update it, you must manually re-run the command.

/cc @bytehead

Commits
-------

3ec5d53c move ide namespace lookup file generation into a service and its own …
fa20d008 add a command to manually dump the ide namespace lookup file
927ead03 write to ""%kernel.build_dir%/contao-ide" by default
25937e0c also pass build dir, when manually warming up in the Automator
8aba7bbe fix default value
5034ccd6 make abbreviation in class names camel case
586d2079 normalize console output when testing strings
be6140ba Also rename the test namespace

Co-authored-by: leofeyer <1192057+leofeyer@users.noreply.github.com>