Always render favorites from the templates (see #9156)
Description
-----------

Fixes #9133

You can insert a Twig template in a PHP Template - we already did that for the message and dialog outlets for instance. I therefore created a single `backend/component/_favorites.html.twig` template that is used anywhere and that abstracts the call to the controller away. 

If you are now using a custom back end controller, favorites will not be generated unless they are output.

Commits
-------

f4c8c7a1 revert #9118 and always render favorites from the templates