Disallow undo preview interaction (see #9169)
Description
-----------

The preview of content elements in the restore/undo view can contain interactive elements, like links, which will not make sense in the back end. For example, the pagination links of a server-side generated pagination of a gallery content element in Contao 5.7 will have links like

```
/contao?do=article&id=105&page_g105=5&table=tl_content&act=delete
```

as they are relative to the current request at the time of deletion of the content element. If you click one of those links you will first get a token error and then an internal server error with 

> Cannot load record "tl_content.id=…".

This PR alleviates that by simply disallowing any interaction with the preview in the undo section.

Commits
-------

efe2fcd2 Disallow undo preview interaction
8b298b62 Revert "Disallow undo preview interaction"
9b098b37 use inert instead